# Create root isolate asynchronously