@liyuqian 2020-08-01
The engine failed to restart in one of our unit tests. Below is some useful error messages that I copied from the [full log]
```
[ RUN      ] ShellTest.MissAtMostOneFrameForIrregularInputEvents
[ERROR:flutter/shell/common/engine.cc(144)] Engine not prepare and launch isolate.
../../flutter/shell/common/shell_test.cc:73: Failure
Value of: restarted.get_future().get()
  Actual: false
Expected: true
```

@scutlight : I'm curious whether you've tested this locally with our `shell_benchmarks`, and what change this PR has brought to the `BM_ShellInitialization` benchmark?

CC @jason-simmons @gaaclarke @chinmaygarde to see if there's any extra concern on deferring loading the isolate.

[full log]: https://logs.chromium.org/logs/flutter/buildbucket/cr-buildbucket.appspot.com/8873295343814216848/+/steps/Host_Tests_for_host_debug_unopt/0/stdout