@scutlight 2020-08-05
@liyuqian : The failed unit test is fixed now. But `shell_benchmarks` runs crashed with current tip-of-tree codes...

> /media/chenjg/1TBSSD/Projects/flutter/engine/src/flutter$ ../out/host_profile/shell_benchmarks 
Run on (8 X 4051.54 MHz CPU s)
2020-08-05 16:13:02
Benchmark                                  Time           CPU Iterations
Segmentation fault (core dumped)

> /media/chenjg/1TBSSD/Projects/flutter/engine/src/flutter$ python -u testing/run_tests.py --type=benchmarks --variant=host_profile --engine-filter=shell_benchmarks
Running Engine Benchmarks.
Running shell_benchmarks in /media/chenjg/1TBSSD/Projects/flutter/engine/src
/media/chenjg/1TBSSD/Projects/flutter/engine/src/out/host_profile/shell_benchmarks
Running command "/media/chenjg/1TBSSD/Projects/flutter/engine/src/out/host_profile/shell_benchmarks"
Run on (8 X 3988.22 MHz CPU s)
2020-08-05 16:13:46
Benchmark                                  Time           CPU Iterations
Traceback (most recent call last):
  File "testing/run_tests.py", line 494, in <module>
    sys.exit(main())
  File "testing/run_tests.py", line 487, in main
    RunEngineBenchmarks(build_dir, engine_filter)
  File "testing/run_tests.py", line 161, in RunEngineBenchmarks
    RunEngineExecutable(build_dir, 'shell_benchmarks', filter)
  File "testing/run_tests.py", line 94, in RunEngineExecutable
    RunCmd(test_command, cwd=cwd)
  File "testing/run_tests.py", line 47, in RunCmd
    raise Exception('Command "%s" exited with code %d' % (command_string, process.returncode))
Exception: Command "/media/chenjg/1TBSSD/Projects/flutter/engine/src/out/host_profile/shell_benchmarks" exited with code -11