@scutlight 2020-08-06
@liyuqian : All failed test cases are resolved now, and benchmarks as follow.
before this pr:
>../out/host_profile/shell_benchmarks 
Run on (8 X 4023.94 MHz CPU s)
2020-08-06 20:39:41
Benchmark                                  Time           CPU Iterations
BM_ShellInitialization               2786435 ns     273327 ns       1000
BM_ShellShutdown                      852115 ns     232808 ns       2974
BM_ShellInitializationAndShutdown   46545353 ns     467147 ns        100

after this pr:
>../out/host_profile/shell_benchmarks 
Run on (8 X 4002.75 MHz CPU s)
2020-08-06 20:21:21
Benchmark                                  Time           CPU Iterations
BM_ShellInitialization                528979 ns     270995 ns       2568
BM_ShellShutdown                      878690 ns     234845 ns       3030
BM_ShellInitializationAndShutdown   46339905 ns     478631 ns        100