@scutlight 2020-08-12
`ShellTest.OnServiceProtocolGetSkSLsWorks` run passed locally but failed on `Cirrus CI`, I'll figure it out later.