@scutlight 2020-08-17
> Mac iOS Engine
> ScenariosTests:
> -[FlutterViewControllerTest testFirstFrameCallback]
> failed, I'll figure it out later.

The failed test case has nothing to do with this pr, just rebase and get all checks passed.